Move utils folder into shared
