+++ packages/runtime/src/widgets/Column.tsx
@@ -4,7 +4,7 @@ import { get } from "lodash-es";
 import { WidgetRegistry } from "../registry";
 import { EnsembleRuntime } from "../runtime";
 import type { FlexboxProps } from "../shared/types";
-import { getColor, getCrossAxis, getMainAxis } from "../util/utils";
+import { getColor, getCrossAxis, getMainAxis } from "../shared/styles";
 
 export const Column: React.FC<FlexboxProps> = (props) => {
   const renderedChildren = useMemo(() => {
