+++ packages/runtime/src/widgets/Image.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import type { Expression } from "@ensembleui/react-framework";
 import { useRegisterBindings } from "@ensembleui/react-framework";
 import { WidgetRegistry } from "../registry";
-import { getColor } from "../util/utils";
+import { getColor } from "../shared/styles";
 import type { EnsembleWidgetProps, HasBorder } from "../shared/types";
 
 export type ImageProps = {
